+++ assets/icon-parts/accent/bonded.png
