+++ assets/icon-parts/frame/purple.png
