+++ docs/scratchpad.md
