+++ scripts/simple-icon.py
