+++ src/services/scratchpad.ts
@@ -10,7 +10,7 @@ import clipboardy from 'clipboardy';
 const ROOT = path.join(path.dirname(fileURLToPath(import.meta.url)), '..', '..');
 
 export function processScratchpad(): string {
-  const file = path.join(ROOT, 'tools', 'scratchpad.md');
+  const file = path.join(ROOT, 'docs', 'scratchpad.md');
   const content = fs.readFileSync(file, 'utf-8');
   const sanitized = content.replace(/\\/g, '\\\\').replace(/"/g, '\\"').replace(/\n/g, '\\n').replace(/\r/g, '');
   clipboardy.writeSync(`"${sanitized}"`);
