+++ tsconfig.node.json
@@ -11,7 +11,9 @@
     "allowImportingTsExtensions": true,
     "verbatimModuleSyntax": true,
     "moduleDetection": "force",
-    "noEmit": true,
+    "composite": true,
+    "emitDeclarationOnly": true,
+    "outDir": "dist/node",
 
     /* Linting */
     "strict": true,
